d3d12: Move/clean code
