+++ rpcs3/Emu/RSX/D3D12/D3D12PipelineState.cpp
@@ -2,7 +2,6 @@
 #if defined (DX12_SUPPORT)
 
 #include "D3D12PipelineState.h"
-#include "D3D12ProgramDisassembler.h"
 #include "Emu/Memory/vm.h"
 #include "Utilities/Log.h"
 #include <wrl/client.h>
