--- rpcs3/Emu/RSX/D3D12/D3D12ProgramDisassembler.h
@@ -1,6 +1,0 @@
-#pragma once
-#if defined (DX12_SUPPORT)
-#include "Emu/RSX/RSXFragmentProgram.h"
-
-void Decompile(RSXFragmentProgram& prog);
-#endif
