+++ rpcs3/Emu/RSX/D3D12/ShaderParam.cpp
@@ -1,0 +1,6 @@
+#include "stdafx.h"
+#if defined(DX12_SUPPORT)
+#include "ShaderParam.h"
+
+
+#endif
