+++ rpcs3/emucore.vcxproj.filters
@@ -884,13 +884,13 @@
     <ClCompile Include="Emu\RSX\D3D12\D3D12PipelineState.cpp">
       <Filter>Emu\GPU\RSX\D3D12</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\D3D12\D3D12ProgramDisassembler.cpp">
-      <Filter>Emu\GPU\RSX\D3D12</Filter>
-    </ClCompile>
     <ClCompile Include="Emu\RSX\D3D12\D3D12Buffer.cpp">
       <Filter>Emu\GPU\RSX\D3D12</Filter>
     </ClCompile>
     <ClCompile Include="Emu\RSX\D3D12\VertexProgramDecompiler.cpp">
+      <Filter>Emu\GPU\RSX\D3D12</Filter>
+    </ClCompile>
+    <ClCompile Include="Emu\RSX\D3D12\ShaderParam.cpp">
       <Filter>Emu\GPU\RSX\D3D12</Filter>
     </ClCompile>
   </ItemGroup>
@@ -1591,13 +1591,13 @@
     <ClInclude Include="Emu\RSX\D3D12\D3D12PipelineState.h">
       <Filter>Emu\GPU\RSX\D3D12</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\D3D12\D3D12ProgramDisassembler.h">
-      <Filter>Emu\GPU\RSX\D3D12</Filter>
-    </ClInclude>
     <ClInclude Include="Emu\RSX\D3D12\D3D12Buffer.h">
       <Filter>Emu\GPU\RSX\D3D12</Filter>
     </ClInclude>
     <ClInclude Include="Emu\RSX\D3D12\VertexProgramDecompiler.h">
+      <Filter>Emu\GPU\RSX\D3D12</Filter>
+    </ClInclude>
+    <ClInclude Include="Emu\RSX\D3D12\ShaderParam.h">
       <Filter>Emu\GPU\RSX\D3D12</Filter>
     </ClInclude>
   </ItemGroup>
